Fix question test issue checking

+++ lib/question.sql
@@ -101,8 +101,8 @@ WHERE
 ORDER BY
    s.date DESC;
 
--- BLOCK select_issues_for_variant
-SELECT COUNT(*)
+-- BLOCK select_issue_count_for_variant
+SELECT COUNT(*)::int
 FROM issues AS i
 WHERE i.variant_id = $variant_id;
